refactor: 불필요한 component 정리
