--- src/Components/Header.js
@@ -1,7 +1,0 @@
-import React, { Component } from "react";
-
-export default class Header extends Component {
-  render() {
-    return <h1>🎱 행운의 로또</h1>;
-  }
-}
